+++ src/OAuth/OAuthLinkCodeGeneratorInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace ThreeBRS\EnterpriseSecurityBundle\OAuth;
+
+interface OAuthLinkCodeGeneratorInterface
+{
+    public function generateCode(): string;
+
+    public function hash(string $plainCode): string;
+}